fix(snap)!: Rename `normalization` mod to `filters`
Inspired by `insta`
